pyproject.toml
[tool.poetry]
name = "habits-service"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
django = "4.2"
djangorestframework = "^3.15.2"
python-dotenv = "^1.0.1"
pillow = "^10.4.0"
djangorestframework-simplejwt = "^5.3.1"
django-cors-headers = "^4.4.0"
pytz = "^2024.1"
celery = "^5.4.0"
redis = "^5.0.8"
django-celery-beat = "^2.7.0"
requests = "^2.32.3"
drf-yasg = "^1.21.7"
psycopg2-binary = "^2.9.9"
coverage = "^7.6.1"
flake8 = "^7.1.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

